feat(spotify): improve scheduling and playlist parsing logic
- Updated `schedule` function to use `int64` for epochMillis, ensuring
  compatibility with `time.UnixMilli`.
- Adjusted delay calculation in `schedule` to use milliseconds for
  higher precision.
- Replaced `strings.SplitN` with `strings.Split` in `parsePlaylistId`
  for better clarity.
- Added unit tests for `schedule` and `parsePlaylistId` to validate
  behavior for various scenarios, including edge cases.

+++ spotify/endpoints.go
@@ -234,7 +234,7 @@ func Schedule(c *gin.Context) {
 		return
 	}
 
-	schedule(epochMillis, fn)
+	schedule(int64(epochMillis), fn)
 	c.JSON(http.StatusOK, gin.H{
 		"message": "Schedule setted successfully",
 	})
